pyproject.toml
[project]
name = "bb-stats"
version = "0.1.3"
description = "Terminal Bitbucket pull request and commit statistics viewer"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0",
  "requests>=2.31",
  "rich>=13.0",
  "textual>=0.80.0",
  "ty",
  "ruff",
]

[project.scripts]
bb-stats = "bb_stats:main"

[build-system]
requires = ["uv_build>=0.11.8,<0.12"]
build-backend = "uv_build"

[tool.uv]
package = true

[tool.uv.build-backend]
module-root = ""

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D101", "D102", "D103", "D107", "CPY001"]
preview = true
